Rename testing module
