+++ p2panda-tests/src/lib.rs
@@ -3,5 +3,5 @@ pub mod utils;
 pub mod logs;
 pub mod node;
 pub mod materializer;
-pub mod testing;
+pub mod data;
 pub mod client;
